refactor: move some things
