+++ packages/react/src/core/forwardRef.ts
@@ -1,10 +1,5 @@
 import { ForwardRefRenderFunction, forwardRef as forwardReactRef } from "react";
-import {
-  RightJoinProps,
-  PropsOf,
-  As,
-  ComponentWithAs,
-} from "../types/extensible.types";
+import { RightJoinProps, PropsOf, As, ComponentWithAs } from "./types";
 
 export function forwardRef<Props extends object, Component extends As>(
   component: ForwardRefRenderFunction<
