--- packages/react/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./extensible.utils";
